+++ Model/Pokemon.py
